Split words files further
This splits the new word-range-class into
word-range-class and to-word, and renames word-range-helpers to
words-helpers.
