+++ Makefile.am
@@ -157,6 +157,8 @@ pkginclude_HEADERS += include/libsemigroups/to-todd-coxeter.hpp
 pkginclude_HEADERS += include/libsemigroups/to-todd-coxeter.tpp
 pkginclude_HEADERS += include/libsemigroups/to-word-graph.hpp
 pkginclude_HEADERS += include/libsemigroups/to-word-graph.tpp
+pkginclude_HEADERS += include/libsemigroups/to-word.hpp
+pkginclude_HEADERS += include/libsemigroups/to-word.tpp
 pkginclude_HEADERS += include/libsemigroups/todd-coxeter-class.hpp
 pkginclude_HEADERS += include/libsemigroups/todd-coxeter-class.tpp
 pkginclude_HEADERS += include/libsemigroups/todd-coxeter-helpers.hpp
@@ -184,9 +186,8 @@ pkginclude_HEADERS += include/libsemigroups/word-graph-view.tpp
 pkginclude_HEADERS += include/libsemigroups/word-graph.hpp
 pkginclude_HEADERS += include/libsemigroups/word-graph.tpp
 pkginclude_HEADERS += include/libsemigroups/word-range-class.hpp
-pkginclude_HEADERS += include/libsemigroups/word-range-class.tpp
-pkginclude_HEADERS += include/libsemigroups/word-range-helpers.hpp
-pkginclude_HEADERS += include/libsemigroups/word-range-helpers.tpp
+pkginclude_HEADERS += include/libsemigroups/words-helpers.hpp
+pkginclude_HEADERS += include/libsemigroups/words-helpers.tpp
 pkginclude_HEADERS += include/libsemigroups/word-range.hpp
 
 detailincludedir = $(includedir)/libsemigroups/detail
@@ -1115,12 +1116,13 @@ libsemigroups_la_SOURCES += src/presentation-examples.cpp
 libsemigroups_la_SOURCES += src/presentation.cpp
 libsemigroups_la_SOURCES += src/runner.cpp
 libsemigroups_la_SOURCES += src/sims.cpp
+libsemigroups_la_SOURCES += src/to-word.cpp
 libsemigroups_la_SOURCES += src/todd-coxeter-helpers.cpp
 libsemigroups_la_SOURCES += src/ukkonen.cpp
 libsemigroups_la_SOURCES += src/word-graph.cpp
 libsemigroups_la_SOURCES += src/word-graph-helpers.cpp
 libsemigroups_la_SOURCES += src/word-range-class.cpp
-libsemigroups_la_SOURCES += src/word-range-helpers.cpp
+libsemigroups_la_SOURCES += src/words-helpers.cpp
 
 libsemigroups_la_SOURCES += src/detail/aho-corasick-impl.cpp
 libsemigroups_la_SOURCES += src/detail/cong-common-class.cpp
