+++ include/libsemigroups/words-helpers.tpp
